+++ src/components/Icons/index.ts
@@ -21,4 +21,6 @@ export { default as ViewMore } from './ViewMore';
 export { default as PublicProfile } from './PublicProfile';
 export { default as Arrow } from './Arrow';
 export { default as Edit } from './Edit';
+export { default as Remove } from './Remove';
+export { default as Plus } from './Plus';
 export { default as Download } from './Download';