+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.13.0] - 2021-06-24
+
+- Added Input component
+- Added Toggle component
+
 ## [1.12.5] - 2021-06-23
 
 - Download icon fixed fill prop
@@ -21,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.12.3] - 2021-06-21
 
 - Public Profile icon reduced to 24px
-- Margin increased to 6px more in tooltip component 
+- Margin increased to 6px more in tooltip component
 
 ## [1.12.2] - 2021-06-18